Corrections in veicles

+++ src/Tanque.java
@@ -1,6 +1,6 @@
 public class Tanque extends Veiculos{
 
-    public Tanque(String nombre, int daño, Float vida) {
+    public Tanque(String nombre, Float daño, Float vida) {
         super(nombre, daño, vida);
     }    
 }
